+++ src/main/java/com/adrien/games/bagl/resource/scene/json/ColorJson.java
@@ -5,6 +5,7 @@ public class ColorJson {
     private float r;
     private float g;
     private float b;
+    private float a;
 
     public float getR() {
         return r;
@@ -16,5 +17,9 @@ public class ColorJson {
 
     public float getB() {
         return b;
+    }
+
+    public float getA() {
+        return a;
     }
 }